requirements.txt
torch==1.9.0+cpu
torchvision==0.10.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
boto3
facenet-pytorch
awslambdaric
requests
pillow
numpy
opencv-python